Update Soundflower link
The old link says:

"Soundflower is an open source kernel extension for Mac OS X, designed to create a virtual audio output device that can also act as an input.

"Initial support for the development and maintenance of Soundflower was provided by Cycling '74. In 2014, Cycling '74 passed stewardship of Soundflower to Rogue Amoeba, and we maintained a fork of the project. Unfortunately, we haven't had the time and resources to devote to improving the product.

"Thankfully, at the end of 2015, Soundflower's original author Matt Ingalls picked up the baton. We expect he'll be providing some updates to Soundflower in the near future, via this GitHub link.

"If you're looking to use Soundflower, that's the link to check out!"